+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2018-08-27
+### Added
+- The methods setWidth and setHeight to Format class.
+### Changed
+- Format.pixelCount is now a getter.
+
 ## 1.0.0 - 2018-08-26
 - Initial release.